backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
requests
pgvector
python-dotenv